pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parol6"
version = "0.1.0"
description = "Python library for controlling PAROL6 robot arms"

requires-python = ">=3.10,<3.12"
dependencies = [
    # TEMPORARY SOLUTION: Using custom-built robotics-toolbox-python wheels from forked repository
    # The proper solution requires the upstream maintainers to complete the merge of their "future" branch.
    # Until then, we use platform-specific wheels from: https://github.com/Jepson2k/robotics-toolbox-python/releases/tag/v1.2.2

    # macOS ARM64 (Apple Silicon)
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp310-cp310-macosx_11_0_arm64.whl ; python_version == '3.10' and platform_system == 'Darwin' and platform_machine == 'arm64'",
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp311-cp311-macosx_11_0_arm64.whl ; python_version == '3.11' and platform_system == 'Darwin' and platform_machine == 'arm64'",

    # macOS x86_64 (Intel)
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp310-cp310-macosx_11_0_x86_64.whl ; python_version == '3.10' and platform_system == 'Darwin' and platform_machine == 'x86_64'",
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp311-cp311-macosx_11_0_x86_64.whl ; python_version == '3.11' and platform_system == 'Darwin' and platform_machine == 'x86_64'",

    # Windows AMD64
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp310-cp310-win_amd64.whl ; python_version == '3.10' and platform_system == 'Windows' and platform_machine == 'AMD64'",
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp311-cp311-win_amd64.whl ; python_version == '3.11' and platform_system == 'Windows' and platform_machine == 'AMD64'",

    # Linux x86_64
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl ; python_version == '3.10' and platform_system == 'Linux' and platform_machine == 'x86_64'",
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp311-cp311-manylinux_2_17_x86_64.manylinux2014_x86_64.whl ; python_version == '3.11' and platform_system == 'Linux' and platform_machine == 'x86_64'",

    # Linux aarch64 (ARM64)
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp310-cp310-manylinux2014_aarch64.manylinux_2_17_aarch64.whl ; python_version == '3.10' and platform_system == 'Linux' and platform_machine == 'aarch64'",
    "roboticstoolbox-python @ https://github.com/Jepson2k/robotics-toolbox-python/releases/download/v1.2.2/roboticstoolbox_python-1.2.2-cp311-cp311-manylinux2014_aarch64.manylinux_2_17_aarch64.whl ; python_version == '3.11' and platform_system == 'Linux' and platform_machine == 'aarch64'",

    "pyserial",
    "spatialmath-python",
    "scipy==1.11.4"
]

[tool.setuptools.packages.find]
include = ["parol6*"]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pre-commit"
]

[project.scripts]
parol6-server = "parol6.cli.server:main_entry"

[tool.pytest.ini_options]
# Limit discovery to the tests/ directory only
testpaths = ["tests"]

# Test discovery patterns
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Output configuration
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "-p", "no:cacheprovider"
]

# Timeout configuration (requires pytest-timeout)
timeout = 300
timeout_method = "thread"

# Logging configuration for tests
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# Pytest markers for different test types
markers = [
    "unit: Unit tests that test individual components in isolation",
    "integration: Integration tests that test component interactions with FAKE_SERIAL",
    "hardware: Hardware tests that require actual robot hardware and human confirmation",
    "slow: Slow-running tests (typically hardware or complex integration tests)",
    "e2e: End-to-end tests that exercise complete workflows",
    "gcode: Tests specifically for GCODE parsing and interpretation functionality"
]

# Filter warnings
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning:roboticstoolbox.*",
    "ignore::UserWarning:spatialmath.*"
]

[[tool.mypy.overrides]]
module = ["roboticstoolbox", "roboticstoolbox.*", "spatialmath", "spatialmath.*"]
ignore_missing_imports = true

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
parol6 = ["urdf_model/**"]

# ---------------- Mypy configuration ----------------
[tool.mypy]
python_version = "3.11"
files = ["parol6"]
exclude = "(?x)(^build/|^dist/|^\\.venv/)"
